services/api-gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.4.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic[email]==2.6.4
python-multipart==0.0.9
pillow==10.2.0
anthropic==0.21.3
google-cloud-vision==3.7.2
scikit-learn==1.4.1
numpy==1.26.4
